Adds Update functionality to bucket
Can click on update button for individual bucket.

Renders modal
Can edit fields in modal

Need to close modal when patch successful
