+++ assets/scripts/templates/bucket.handlebars
@@ -17,6 +17,6 @@
   </div>
   <div class="bucketEditDeleteButtons">
     <button type="submit" class="delete-bucket-button btn btn-default" id="{{@index}}" data-id="{{_id}}">Delete</button>
-    <button type="submit" class="edit-bucket-button btn btn-default" id="{{@index}}" data-id="{{_id}}">Update</button>
+    <button type="submit" class="edit-bucket-button btn btn-default" id="{{@index}}" data-id="{{_id}}" data-toggle="modal" data-target="#modal-fullscreen-update-bucket">Update</button>
   </div>
 </div>
