+++ assets/scripts/templates/update-bucket.handlebars
@@ -1,0 +1,2 @@
+
+  {{> update-bucket-form}}
