+++ index.html
@@ -165,6 +165,27 @@
       </div>
     </div>
 
+
+    <!-- Modal fullscreen Update Bucket -->
+    <div class="modal modal-fullscreen fade update-bucket" id="modal-fullscreen-update-bucket" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
+      <div class="modal-dialog">
+        <div class="modal-content">
+          <div class="modal-header">
+            <button type="button" class="close" data-dismiss="modal"><span aria-hidden="true">&times;</span><span class="sr-only">Close</span></button>
+            <h4 class="modal-title" id="myModalLabel">Edit Your Bucket</h4>
+          </div>
+          <div class="modal-body">
+            <div id="update-bucket-form">
+            </div>
+          </div>
+          <div class="modal-footer">
+            <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
+          </div>
+        </div>
+      </div>
+    </div>
+
+
     <!-- CREATE PROFILE FORM -->
     <section class="create-profile">
       <form id="create-profile">
